day12 project/pyproject.toml
[tool.poetry]
name = "best-restaurant"
version = "0.1.0"
description = ""
authors = ["meoyoug <eodyd426@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-extensions = "^3.2.3"
mysqlclient = "^2.2.4"
pillow = "^10.4.0"
django-cleanup = "^8.1.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
